Handle collective.indexing and Products.remember more nicely.
Create a bin/test-indexing script that had collective.indexing in it.

Products.remember 1.9.4 does not work on Plone 5, so I split that out.
We now have a remember.cfg buildout config that creates a bin/test-remember script.
If the Products.remember package is available, we import it and load its zcml in the tests.