Split user and group groups interface.
A group that implemented ``user.IUserGroupsProvider`` would get
included when listing members, which led to ``AttributeError:
'NoneType' object has no attribute '__of__'``.  The new interface is
``group.IGroupGroupsProvider``.  We look for this in our groupmanager
plugin in the ``getGroupsForPrincipal`` method.  If no providers are
found, we try the old way for backwards compatibility.